requirements.txt
flask==2.0.1
flask-dance==5.0.0
python-decouple==3.4
gunicorn==20.1.0
pyopenssl===21.0.0
